+++ projects/packages/forms/src/blocks/contact-form/child-blocks.js
@@ -2,7 +2,7 @@ import { InnerBlocks } from '@wordpress/block-editor';
 import { createBlock } from '@wordpress/blocks';
 import { Path, Icon } from '@wordpress/components';
 import { __, _x } from '@wordpress/i18n';
-import { globe, envelope, mobile } from '@wordpress/icons';
+import { globe, envelope, mobile, upload } from '@wordpress/icons';
 import { filter, isEmpty, map, startsWith, trim } from 'lodash';
 import JetpackField from './components/jetpack-field';
 import JetpackFieldCheckbox from './components/jetpack-field-checkbox';
@@ -31,6 +31,7 @@ const FieldDefaults = {
 		label: {
 			type: 'string',
 			default: null,
+			role: 'content',
 		},
 		required: {
 			type: 'boolean',
@@ -38,18 +39,22 @@ const FieldDefaults = {
 		},
 		requiredText: {
 			type: 'string',
+			role: 'content',
 		},
 		options: {
 			type: 'array',
 			default: [],
+			role: 'content',
 		},
 		defaultValue: {
 			type: 'string',
 			default: '',
+			role: 'content',
 		},
 		placeholder: {
 			type: 'string',
 			default: '',
+			role: 'content',
 		},
 		id: {
 			type: 'string',
@@ -233,12 +238,16 @@ const FieldDefaults = {
 	example: {},
 };
 
+// Storing in variables to avoid JS mangling breaking translation calls
+const severalOptionsDefault = __( 'Choose several options', 'jetpack-forms' );
+const oneOptionDefault = __( 'Choose one option', 'jetpack-forms' );
+
 const multiFieldV1 = fieldType => ( {
 	attributes: {
 		...FieldDefaults.attributes,
 		label: {
 			type: 'string',
-			default: fieldType === 'checkbox' ? 'Choose several options' : 'Choose one option',
+			default: fieldType === 'checkbox' ? severalOptionsDefault : oneOptionDefault,
 		},
 	},
 	migrate: attributes => {
@@ -365,7 +374,8 @@ export const childBlocks = [
 				...FieldDefaults.attributes,
 				label: {
 					type: 'string',
-					default: 'Text',
+					default: __( 'Text', 'jetpack-forms' ),
+					role: 'content',
 				},
 			},
 		},
@@ -375,7 +385,7 @@ export const childBlocks = [
 		settings: {
 			...FieldDefaults,
 			title: __( 'Name Field', 'jetpack-forms' ),
-			description: __( 'Collect the site visitor’s name.', 'jetpack-forms' ),
+			description: __( "Collect the site visitor's name.", 'jetpack-forms' ),
 			icon: {
 				foreground: getIconColor(),
 				src: renderMaterialIcon(
@@ -387,7 +397,8 @@ export const childBlocks = [
 				...FieldDefaults.attributes,
 				label: {
 					type: 'string',
-					default: 'Name',
+					default: __( 'Name', 'jetpack-forms' ),
+					role: 'content',
 				},
 			},
 		},
@@ -408,7 +419,8 @@ export const childBlocks = [
 				...FieldDefaults.attributes,
 				label: {
 					type: 'string',
-					default: 'Email',
+					default: __( 'Email', 'jetpack-forms' ),
+					role: 'content',
 				},
 			},
 		},
@@ -435,6 +447,7 @@ export const childBlocks = [
 				label: {
 					type: 'string',
 					default: __( 'Website', 'jetpack-forms' ),
+					role: 'content',
 				},
 			},
 		},
@@ -463,7 +476,8 @@ export const childBlocks = [
 				...FieldDefaults.attributes,
 				label: {
 					type: 'string',
-					default: 'Date',
+					default: __( 'Date', 'jetpack-forms' ),
+					role: 'content',
 				},
 				dateFormat: {
 					type: 'string',
@@ -493,8 +507,40 @@ export const childBlocks = [
 				label: {
 					type: 'string',
 					default: 'Phone',
+					role: 'content',
 				},
 			},
+		},
+	},
+	{
+		name: 'field-file',
+		settings: {
+			...FieldDefaults,
+			title: __( 'File Upload Field', 'jetpack-forms' ),
+			keywords: [
+				__( 'File', 'jetpack-forms' ),
+				__( 'Upload', 'jetpack-forms' ),
+				__( 'Attachment', 'jetpack-forms' ),
+			],
+			description: __( 'Allow visitors to upload files through your form.', 'jetpack-forms' ),
+			icon: {
+				foreground: getIconColor(),
+				src: <Icon icon={ upload } />,
+			},
+			edit: editField( 'file' ),
+			attributes: {
+				...FieldDefaults.attributes,
+				label: {
+					type: 'string',
+					default: __( 'Upload a file', 'jetpack-forms' ),
+					role: 'content',
+				},
+				filetype: {
+					type: 'string',
+					default: '',
+				},
+			},
+			isBeta: true,
 		},
 	},
 	{
@@ -542,6 +588,7 @@ export const childBlocks = [
 				label: {
 					type: 'string',
 					default: '',
+					role: 'content',
 				},
 			},
 		},
@@ -649,10 +696,12 @@ export const childBlocks = [
 				toggleLabel: {
 					type: 'string',
 					default: null,
+					role: 'content',
 				},
 				options: {
 					type: 'array',
 					default: [ '' ],
+					role: 'content',
 				},
 			},
 		},