+++ projects/packages/videopress/CHANGELOG.md
@@ -5,9 +5,28 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.25.11] - 2025-02-11
+### Changed
+- Update dependencies. [#39855]
+
+## [0.25.10] - 2025-02-10
+### Changed
+- Updated package dependencies. [#41491]
+- Updated package dependencies. [#41577]
+
+### Fixed
+- VideoPress: Fix warnnings for videos with no prior metadata. [#41639]
+
+## [0.25.9] - 2025-02-03
+### Changed
+- Updated package dependencies. [#41286]
+
+### Fixed
+- Code: Remove extra params on function calls. [#41263]
+
 ## [0.25.8] - 2025-01-27
 ### Fixed
-- VideoPress: Fix issue with VideoPress block with 0 height and width [#41319]
+- VideoPress: Fix issue with VideoPress block with 0 height and width. [#41319]
 
 ## [0.25.7] - 2025-01-20
 ### Changed
@@ -1521,6 +1540,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Created empty package [#24952]
 
+[0.25.11]: https://github.com/Automattic/jetpack-videopress/compare/v0.25.10...v0.25.11
+[0.25.10]: https://github.com/Automattic/jetpack-videopress/compare/v0.25.9...v0.25.10
+[0.25.9]: https://github.com/Automattic/jetpack-videopress/compare/v0.25.8...v0.25.9
 [0.25.8]: https://github.com/Automattic/jetpack-videopress/compare/v0.25.7...v0.25.8
 [0.25.7]: https://github.com/Automattic/jetpack-videopress/compare/v0.25.6...v0.25.7
 [0.25.6]: https://github.com/Automattic/jetpack-videopress/compare/v0.25.5...v0.25.6